Rough Draft of Shiny App
TL;DR I have a Shiny app I'm developing that would produce a ggplot histogram (not shown) from a series of Excel sheets the user uploads. The problem is getting a list of colnames so the user can choose which columns to use for the histogram.

I have come up with the following code from various useful questions on Stack Exchange and GitHub, but for some reason, when I run it, Shiny does not produce a SelectInput for the Excel sheets like I want it to. The finished code should produce a unique SelectInput for each Excel sheet uploaded, and the user firstly sets how many excel sheets will be taken.

I think the problem is in my observers and reactives, though it likely could also be in my call to HTML. Please see below and let me know how I can make this work. Apologies if this code is messy; it's my very first Shiny app.